Use functions to analyze data and represent it
The "read.cvs" function was first used to bring the data file into R Studio. Then, the data was plotted using the plot function. An nls function was crafted afterwards to represent a model for the data, as it represented a sine wave. Curve and Predict were used to farther analzye the data and confirm whether it adhered to the given nls model.